requirements.txt
Django==4.2.28
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
PyYAML==6.0.3
pytz==2025.2